N64: Rename n64_mempak files
